refactor: move tab icon to components
