+++ app/(tabs)/_layout.tsx
@@ -1,47 +1,7 @@
-import { View, Image, ImageURISource, Text } from 'react-native';
 import React from 'react';
 import { Tabs } from 'expo-router';
 import icons from '@/constants/icons';
-
-type TabIconProps = {
-    icon: ImageURISource;
-    color: string;
-    name: string;
-    focused?: boolean;
-};
-
-const TabIcon = ({ icon, color, name }: TabIconProps) => {
-    return (
-        <View
-            style={{
-                display: 'flex',
-                alignItems: 'center',
-            }}
-        >
-            <Image
-                source={icon}
-                resizeMode="contain"
-                tintColor={color}
-                style={{
-                    width: 28,
-                    height: 28,
-                    marginBottom: 5,
-                }}
-            />
-            <Text
-                style={{
-                    fontFamily: 'Work Sans',
-                    fontSize: 12,
-                    fontWeight: 500,
-                    lineHeight: 16,
-                    textAlign: 'center',
-                }}
-            >
-                {name}
-            </Text>
-        </View>
-    );
-};
+import { TabIcon } from '@/components/TabIcon/TabIcon';
 
 const TabsLayout = () => {
     return (
